requirements.txt
# LocalBooru Python dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Image processing
pillow>=10.0.0
imagehash>=4.3.0

# ONNX Runtime for tagger
onnxruntime>=1.18.0
# Use onnxruntime-gpu for CUDA support:
# onnxruntime-gpu>=1.18.0

# Face detection and age estimation
insightface>=0.7.3
opencv-python-headless>=4.8.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0

# Filesystem watching
watchdog>=4.0.0

# UPnP for network port forwarding
miniupnpc>=2.2.0
